refactor: remove nested components in ConfigPanel
